refactor(server): Move routes into seperate files
