+++ server/src/router/shared.ts
@@ -1,0 +1,4 @@
+import { config } from '../config'
+import { getEventDatabases } from '../dbUtils'
+
+export let { event, eventData } = getEventDatabases(config.eventId)
